backend/requirements.txt
# FastAPI Backend Requirements
# Spec 4: Frontend-Backend Integration

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# CORS support
python-multipart>=0.0.9

# Environment variables
python-dotenv>=1.0.0

# Data validation (included with FastAPI)
pydantic>=2.0.0

# RAG Agent dependencies (from Spec 3)
# Install separately: pip install -r ../rag-pipeline/requirements-agent.txt
